requirements.txt
# CrowdCode Platform Dependencies

# GitHub API
PyGithub==2.1.1

# YAML parsing
PyYAML==6.0.1

# Future: AI code generation
# openai>=1.0.0
